+++ packages/remax-cli/src/__tests__/integration/renamePage.test.ts
@@ -1,0 +1,5 @@
+import runTest from './helpers/runTest';
+
+describe('rename Page identifier', () => {
+  runTest('renamePage');
+});