+++ packages/remax-one/src/components/Modal.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { usePageContext, createPortal } from '@remax/runtime';
 
-export default function Modal({ children }: React.PropsWithChildren<{}>) {
+export default function Modal({ children }: React.PropsWithChildren<Record<string, any>>) {
   // context 一定存在
   const ctx = usePageContext()!;