+++ packages/remax-wechat/src/hostComponents/Navigator/index.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { createHostComponent } from '@remax/shared';
-import { BaseProps } from '../../types/component';
+import { BaseProps, GenericEvent } from '../../types/component';
 
 export interface NavigatorProps extends BaseProps {
   /** (default: self) 在哪个目标上发生跳转，默认当前小程序 2.0.7 */
@@ -16,7 +16,7 @@ export interface NavigatorProps extends BaseProps {
   /** 当target="miniProgram"时有效，打开的页面路径，如果为空则打开首页 2.0.7 */
   path?: string;
   /** 当target="miniProgram"时有效，需要传递给目标小程序的数据，目标小程序可在 App.onLaunch()，App.onShow() 中获取到这份数据。详情 2.0.7 */
-  extraData?: object;
+  extraData?: any;
   /** (default: release) 当target="miniProgram"时有效，要打开的小程序版本 2.0.7 */
   version?: 'develop' | 'trial' | 'release';
   /** (default: navigator-hover) 指定点击时的样式类，当hover-class="none"时，没有点击态效果 1.0.0 */
@@ -28,11 +28,11 @@ export interface NavigatorProps extends BaseProps {
   /** (default: 600) 手指松开后点击态保留时间，单位毫秒 1.0.0 */
   hoverStayTime?: number;
   /** 当target="miniProgram"时有效，跳转小程序成功 2.0.7 */
-  onSuccess?: (event: any) => any;
+  onSuccess?: (event: GenericEvent) => any;
   /** 当target="miniProgram"时有效，跳转小程序失败 2.0.7 */
-  onFail?: (event: any) => any;
+  onFail?: (event: GenericEvent) => any;
   /** 当target="miniProgram"时有效，跳转小程序完成 2.0.7 */
-  onComplete?: (event: any) => any;
+  onComplete?: (event: GenericEvent) => any;
 }
 /**
  * https://developers.weixin.qq.com/miniprogram/dev/component/navigator.html