+++ packages/remax-wechat/src/hostComponents/OpenData/index.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { createHostComponent } from '@remax/shared';
-import { BaseProps } from '../../types/component';
+import { BaseProps, GenericEvent } from '../../types/component';
 
 export interface OpenDataProps extends BaseProps {
   /** 开放数据类型 1.4.0 */
@@ -22,7 +22,7 @@ export interface OpenDataProps extends BaseProps {
   /** 用户头像为空时的默认图片，支持相对路径和网络图片路径 2.8.1 */
   defaultAvatar?: string;
   /** 群名称或用户信息为空时触发 2.8.1 */
-  onError?: (event: any) => any;
+  onError?: (event: GenericEvent) => any;
 }
 /**
  * https://developers.weixin.qq.com/miniprogram/dev/component/open-data.html