+++ packages/remax-wechat/src/hostComponents/Picker/index.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { createHostComponent } from '@remax/shared';
-import { BaseProps } from '../../types/component';
+import { BaseProps, GenericEvent } from '../../types/component';
 
 export interface PickerProps extends BaseProps {
   name?: string;
@@ -24,8 +24,8 @@ export interface PickerProps extends BaseProps {
    * 1.9.90
    * 取消选择时触发
    */
-  onCancel?: (event: any) => any;
-  onChange?: (event: any) => void;
+  onCancel?: (event: GenericEvent) => any;
+  onChange?: (event: GenericEvent) => void;
   /**
    * 列改变时触发
    */
@@ -65,11 +65,6 @@ export interface PickerProps extends BaseProps {
    * 可为每一列的顶部添加一个自定义的项
    */
   customItem?: string;
-  /**
-   * 点击时触发
-   */
-  onClick?: (event: any) => any;
-  onTap?: (event: any) => any;
 }
 
 /**