+++ packages/remax-wechat/src/types/index.ts
@@ -1,1 +1,2 @@
 export * from './config';
+export { TouchEvent, GenericEvent } from './component';